Move all user stuff into the final location
