+++ politeiawww/paywall.go
@@ -53,129 +53,6 @@ func (p *politeiawww) paywallIsEnabled() bool {
 	return p.cfg.PaywallAmount != 0 && p.cfg.PaywallXpub != ""
 }
 
-// addUserToPaywallPool adds a database user to the paywall pool.
-//
-// This function must be called WITH the mutex held.
-func (p *politeiawww) addUserToPaywallPool(u *user.User, paywallType string) {
-	p.userPaywallPool[u.ID] = paywallPoolMember{
-		paywallType: paywallType,
-		address:     u.NewUserPaywallAddress,
-		amount:      u.NewUserPaywallAmount,
-		txNotBefore: u.NewUserPaywallTxNotBefore,
-		pollExpiry:  u.NewUserPaywallPollExpiry,
-	}
-}
-
-// addUserToPaywallPoolLock adds a user and its paywall info to the in-memory pool.
-//
-// This function must be called WITHOUT the mutex held.
-func (p *politeiawww) addUserToPaywallPoolLock(u *user.User, paywallType string) {
-	if !p.paywallIsEnabled() {
-		return
-	}
-
-	p.Lock()
-	defer p.Unlock()
-
-	p.addUserToPaywallPool(u, paywallType)
-}
-
-func (p *politeiawww) updateUserAsPaid(u *user.User, tx string) error {
-	u.NewUserPaywallTx = tx
-	u.NewUserPaywallPollExpiry = 0
-	return p.db.UserUpdate(*u)
-}
-
-func (p *politeiawww) derivePaywallInfo(u *user.User) (string, uint64, int64, error) {
-	address, err := util.DerivePaywallAddress(p.params,
-		p.cfg.PaywallXpub, uint32(u.PaywallAddressIndex))
-	if err != nil {
-		err = fmt.Errorf("Unable to derive paywall address #%v "+
-			"for %v: %v", u.ID.ID(), u.Email, err)
-	}
-
-	return address, p.cfg.PaywallAmount, time.Now().Unix(), err
-}
-
-func (p *politeiawww) createUserPaywallPoolCopy() map[uuid.UUID]paywallPoolMember {
-	p.RLock()
-	defer p.RUnlock()
-
-	poolCopy := make(map[uuid.UUID]paywallPoolMember, len(p.userPaywallPool))
-
-	for k, v := range p.userPaywallPool {
-		poolCopy[k] = v
-	}
-
-	return poolCopy
-}
-
-func (p *politeiawww) checkForUserPayments(pool map[uuid.UUID]paywallPoolMember) (bool, []uuid.UUID) {
-	var userIDsToRemove []uuid.UUID
-
-	for userID, poolMember := range pool {
-		u, err := p.db.UserGetById(userID)
-		if err != nil {
-			if err == user.ErrShutdown {
-				// The database is shutdown, so stop the thread.
-				return false, nil
-			}
-
-			log.Errorf("cannot fetch user by id %v: %v\n", userID, err)
-			continue
-		}
-
-		if poolMember.paywallType != paywallTypeUser {
-			continue
-		}
-
-		log.Tracef("Checking the user paywall address for user %v...", u.Email)
-
-		if p.HasUserPaid(u) {
-			// The user could have been marked as paid by RouteVerifyUserPayment,
-			// so just remove him from the in-memory pool.
-			userIDsToRemove = append(userIDsToRemove, userID)
-			log.Tracef("  removing from polling, user already paid")
-			continue
-		}
-
-		if paywallHasExpired(u.NewUserPaywallPollExpiry) {
-			userIDsToRemove = append(userIDsToRemove, userID)
-			log.Tracef("  removing from polling, poll has expired")
-			continue
-		}
-
-		tx, _, err := util.FetchTxWithBlockExplorers(poolMember.address, poolMember.amount,
-			poolMember.txNotBefore, p.cfg.MinConfirmationsRequired)
-		if err != nil {
-			log.Errorf("cannot fetch tx: %v\n", err)
-			continue
-		}
-
-		if tx != "" {
-			// Update the user in the database.
-			err = p.updateUserAsPaid(u, tx)
-			if err != nil {
-				if err == user.ErrShutdown {
-					// The database is shutdown, so stop the thread.
-					return false, nil
-				}
-
-				log.Errorf("cannot update user with id %v: %v", u.ID, err)
-				continue
-			}
-
-			// Remove this user from the in-memory pool.
-			userIDsToRemove = append(userIDsToRemove, userID)
-			log.Tracef("  removing from polling, user just paid")
-		}
-
-		time.Sleep(paywallCheckGap)
-	}
-
-	return true, userIDsToRemove
-}
-
 // checkForProposalPayments checks if any of the proposal paywalls in the
 // paywall pool have received a payment.  If so, proposal credits are created
 // for the user, the user database is updated, and the user is removed from
@@ -246,15 +123,6 @@ func (p *politeiawww) checkForProposalPayments(pool map[uuid.UUID]paywallPoolMem
 	return true, userIDsToRemove
 }
 
-func (p *politeiawww) removeUsersFromPool(userIDsToRemove []uuid.UUID) {
-	p.Lock()
-	defer p.Unlock()
-
-	for _, userID := range userIDsToRemove {
-		delete(p.userPaywallPool, userID)
-	}
-}
-
 func (p *politeiawww) checkForPayments() {
 	for {
 		// Removing pool members from the pool while in the middle of
@@ -278,154 +146,6 @@ func (p *politeiawww) checkForPayments() {
 
 		time.Sleep(paywallCheckGap)
 	}
-}
-
-// GenerateNewUserPaywall generates new paywall info, if necessary, and saves
-// it in the database.
-func (p *politeiawww) GenerateNewUserPaywall(u *user.User) error {
-	// Check that the paywall is enabled.
-	if !p.paywallIsEnabled() {
-		return nil
-	}
-
-	// Check that the user either hasn't had paywall information set yet,
-	// or it has expired.
-	if u.NewUserPaywallAddress != "" &&
-		!paywallHasExpired(u.NewUserPaywallPollExpiry) {
-		return nil
-	}
-
-	if u.NewUserPaywallAddress == "" {
-		address, amount, txNotBefore, err := p.derivePaywallInfo(u)
-		if err != nil {
-			return err
-		}
-
-		u.NewUserPaywallAddress = address
-		u.NewUserPaywallAmount = amount
-		u.NewUserPaywallTxNotBefore = txNotBefore
-	}
-	u.NewUserPaywallPollExpiry = time.Now().Add(paywallExpiryDuration).Unix()
-
-	err := p.db.UserUpdate(*u)
-	if err != nil {
-		return err
-	}
-
-	p.addUserToPaywallPoolLock(u, paywallTypeUser)
-	return nil
-}
-
-// ProcessVerifyUserPayment verifies that the provided transaction
-// meets the minimum requirements to mark the user as paid, and then does
-// that in the user database.
-func (p *politeiawww) ProcessVerifyUserPayment(u *user.User, vupt v1.VerifyUserPayment) (*v1.VerifyUserPaymentReply, error) {
-	var reply v1.VerifyUserPaymentReply
-	if p.HasUserPaid(u) {
-		reply.HasPaid = true
-		return &reply, nil
-	}
-
-	if paywallHasExpired(u.NewUserPaywallPollExpiry) {
-		err := p.GenerateNewUserPaywall(u)
-		if err != nil {
-			return nil, err
-		}
-		reply.PaywallAddress = u.NewUserPaywallAddress
-		reply.PaywallAmount = u.NewUserPaywallAmount
-		reply.PaywallTxNotBefore = u.NewUserPaywallTxNotBefore
-		return &reply, nil
-	}
-
-	tx, _, err := util.FetchTxWithBlockExplorers(u.NewUserPaywallAddress,
-		u.NewUserPaywallAmount, u.NewUserPaywallTxNotBefore,
-		p.cfg.MinConfirmationsRequired)
-	if err != nil {
-		if err == util.ErrCannotVerifyPayment {
-			return nil, v1.UserError{
-				ErrorCode: v1.ErrorStatusCannotVerifyPayment,
-			}
-		}
-		return nil, err
-	}
-
-	if tx != "" {
-		reply.HasPaid = true
-
-		err = p.updateUserAsPaid(u, tx)
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		// TODO: Add the user to the in-memory pool.
-	}
-
-	return &reply, nil
-}
-
-// HasUserPaid checks that a user has paid the paywall
-func (p *politeiawww) HasUserPaid(u *user.User) bool {
-	// Return true when running unit tests
-	if p.test {
-		return true
-	}
-
-	// Return true if paywall is disabled
-	if !p.paywallIsEnabled() {
-		return true
-	}
-
-	return u.NewUserPaywallTx != ""
-}
-
-func (p *politeiawww) addUsersToPaywallPool() error {
-	p.Lock()
-	defer p.Unlock()
-
-	// Create the in-memory pool of all users who need to pay the paywall.
-	err := p.db.AllUsers(func(u *user.User) {
-		// Proposal paywalls
-		if p.userHasValidProposalPaywall(u) {
-			p.addUserToPaywallPool(u, paywallTypeProposal)
-			return
-		}
-
-		// User paywalls
-		if p.HasUserPaid(u) {
-			return
-		}
-		if u.NewUserVerificationToken != nil {
-			return
-		}
-		if paywallHasExpired(u.NewUserPaywallPollExpiry) {
-			return
-		}
-
-		p.addUserToPaywallPool(u, paywallTypeUser)
-	})
-	if err != nil {
-		return err
-	}
-
-	log.Tracef("Adding %v users to paywall pool", len(p.userPaywallPool))
-	return nil
-}
-
-// initPaywallCheck is intended to be called
-func (p *politeiawww) initPaywallChecker() error {
-	if p.cfg.PaywallAmount == 0 {
-		// Paywall not configured.
-		return nil
-	}
-
-	err := p.addUsersToPaywallPool()
-	if err != nil {
-		return err
-	}
-
-	// Start the thread that checks for payments.
-	go p.checkForPayments()
-	return nil
 }
 
 // mostRecentProposalPaywall returns the most recent paywall that has been
